7월19일 오전11시50분 itemForm.html 상품등록 수정중 main.html 메인페이지 무한스크롤 적용됨
오후6시20분
main.html 수정

8월19일 오후2시30분
최신버전, discountService 추가

+++ src/main/java/com/shop/controller/ThymeleafExController.java
@@ -8,6 +8,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -28,7 +29,7 @@ public class ThymeleafExController {
         ItemDto itemDto = new ItemDto();
         itemDto.setItemDetail("상품 상세 설명");
         itemDto.setItemNm("테스트 상품1");
-        itemDto.setPrice(10000);
+        itemDto.setPrice(BigDecimal.valueOf(10000));  // 숫자를 BigDecimal로 변환하여 설정 8월19일
         itemDto.setRegTime(LocalDateTime.now());
         model.addAttribute("itemDto",itemDto);
         return "thymeleafEx/thymeleafEx02";
@@ -39,7 +40,7 @@ public class ThymeleafExController {
         for(int i = 1;i<=10;i++){
             ItemDto item = new ItemDto();
             item.setItemNm("테스트 상품" + i);
-            item.setPrice(10000 + i);
+            item.setPrice(BigDecimal.valueOf(10000 + i)); // BigDecimal로 표현식 변경 8월19일
             item.setItemDetail("테스트 상품 상세 설명" + i);
             item.setSellStatCd("SELL");
             item.setRegTime(LocalDateTime.now());
@@ -55,7 +56,7 @@ public class ThymeleafExController {
         for(int i = 1;i<=10;i++){
             ItemDto item = new ItemDto();
             item.setItemNm("테스트 상품" + i);
-            item.setPrice(10000 + i);
+            item.setPrice(BigDecimal.valueOf(10000 + i)); // BigDecimal로 표현식 변경 8월19일
             item.setItemDetail("테스트 상품 상세 설명" + i);
             item.setSellStatCd("SELL");
             item.setRegTime(LocalDateTime.now());
