+++ src/main/java/com/shop/dto/ItemDto.java
@@ -3,6 +3,7 @@ package com.shop.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -10,7 +11,7 @@ import java.time.LocalDateTime;
 public class ItemDto {
     private Long id;
     private String itemNm;
-    private Integer price;
+    private BigDecimal price; // 8월19일 수정
     private String itemDetail;
     private String sellStatCd;
     private LocalDateTime regTime;
