+++ src/main/java/com/shop/dto/MainItemDto.java
@@ -4,6 +4,8 @@ import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class MainItemDto {
@@ -11,10 +13,10 @@ public class MainItemDto {
     private String itemNm;
     private String itemDetail;
     private String imgUrl;
-    private Integer price;
+    private BigDecimal price; // int -> BigDecimal 변경 8월19일
     private int stockNumber;
     @QueryProjection //Querydsl 결과 조회 시 MainItemDto 객체로 바로 오도록  활용
-    public MainItemDto(Long id, String itemNm, String itemDetail, String imgUrl, Integer price, int stockNumber){
+    public MainItemDto(Long id, String itemNm, String itemDetail, String imgUrl, BigDecimal price, int stockNumber){ // int -> BigDecimal 변경 8월19일
         this.id = id;
         this.itemNm = itemNm;
         this.itemDetail = itemDetail;
