+++ src/main/java/com/shop/dto/OrderItemDto.java
@@ -4,12 +4,14 @@ import com.shop.entity.OrderItem;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class OrderItemDto {
     private String itemNm;
     private int count;
-    private int orderPrice;
+    private BigDecimal orderPrice; // int -> BigDecimal 변경 8월19일
     private String imgUrl;
     public OrderItemDto(OrderItem orderItem, String imgUrl){
         this.itemNm = orderItem.getItem().getItemNm();
